+++ src/main/java/com/eventra/eventnotifiation/model/EventNotifiationRepository.java
@@ -1,4 +1,4 @@
-package com.eventra.event_notifiation.model;
+package com.eventra.eventnotifiation.model;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -6,8 +6,6 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.transaction.annotation.Transactional;
-
-import com.eventra.order_notification.model.OrderNotificationVO;
 
 public interface EventNotifiationRepository extends JpaRepository<EventNotifiationVO, Integer> {
